+++ src/main/java/kitchenpos/order/domain/NumberOfGuests.java
@@ -1,4 +1,4 @@
-package kitchenpos.ordertable.domain;
+package kitchenpos.order.domain;
 
 import kitchenpos.common.constant.ErrorCode;
