+++ src/main/java/kitchenpos/order/domain/OrderTable.java
@@ -1,7 +1,4 @@
-package kitchenpos.ordertable.domain;
-
-import kitchenpos.common.constant.ErrorCode;
-import kitchenpos.order.domain.OrderStatus;
+package kitchenpos.order.domain;
 
 import javax.persistence.*;
 import java.util.Objects;
@@ -15,8 +12,6 @@ public class OrderTable {
     private NumberOfGuests numberOfGuests;
     private boolean empty;
     private Long tableGroupId;
-    @Enumerated(EnumType.STRING)
-    private OrderStatus orderStatus;
 
     protected OrderTable() {}
