+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupRequest.java
@@ -1,7 +1,5 @@
 package kitchenpos.tablegroup.dto;
 
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.domain.OrderTables;
 import kitchenpos.tablegroup.domain.TableGroup;
 
 import java.time.LocalDateTime;
