+++ src/test/java/kitchenpos/order/domain/OrderTest.java
@@ -2,8 +2,6 @@ package kitchenpos.order.domain;
 
 import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.fixture.TestOrderFactory;
-import kitchenpos.ordertable.domain.NumberOfGuests;
-import kitchenpos.ordertable.domain.OrderTable;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
