+++ src/test/java/kitchenpos/ordertable/application/TableServiceTest.java
@@ -1,17 +1,14 @@
 package kitchenpos.ordertable.application;
 
-import kitchenpos.fixture.TestOrderFactory;
-import kitchenpos.order.domain.Order;
-import kitchenpos.order.domain.OrderStatus;
-import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.ordertable.domain.NumberOfGuests;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.dto.OrderTableRequest;
-import kitchenpos.ordertable.dto.OrderTableResponse;
-import kitchenpos.ordertable.dto.UpdateEmptyRequest;
-import kitchenpos.ordertable.dto.UpdateNumberOfGuestsRequest;
-import kitchenpos.ordertable.repository.OrderTableRepository;
-import kitchenpos.ordertable.validator.OrderTableValidator;
+import kitchenpos.order.application.TableService;
+import kitchenpos.order.domain.NumberOfGuests;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.dto.OrderTableRequest;
+import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.order.dto.UpdateEmptyRequest;
+import kitchenpos.order.dto.UpdateNumberOfGuestsRequest;
+import kitchenpos.order.repository.OrderTableRepository;
+import kitchenpos.order.validator.OrderTableValidator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -19,7 +16,6 @@ import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
