+++ src/test/java/kitchenpos/ordertable/domain/OrderTableTest.java
@@ -1,16 +1,16 @@
 package kitchenpos.ordertable.domain;
 
-import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.fixture.TestOrderFactory;
+import kitchenpos.order.domain.NumberOfGuests;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderStatus;
+import kitchenpos.order.domain.OrderTable;
 import kitchenpos.tablegroup.domain.TableGroup;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
 import java.util.ArrayList;
-import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
