+++ src/test/java/kitchenpos/tablegroup/validator/TableGroupValidatorTest.java
@@ -4,9 +4,9 @@ import kitchenpos.common.constant.ErrorCode;
 import kitchenpos.fixture.TestOrderFactory;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.repository.OrderRepository;
-import kitchenpos.ordertable.domain.NumberOfGuests;
-import kitchenpos.ordertable.domain.OrderTable;
-import kitchenpos.ordertable.repository.OrderTableRepository;
+import kitchenpos.order.domain.NumberOfGuests;
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.repository.OrderTableRepository;
 import kitchenpos.tablegroup.domain.TableGroup;
 import kitchenpos.tablegroup.dto.TableGroupRequest;
 import org.junit.jupiter.api.DisplayName;
